Reorder a GarbageCollector::RunFinalizer signature to fix matching on SVS
The "Test Game" signature is a bit general that it matched `MS.Internal.Xml.Cache.XPathDocumentNavigator` in Summer Vacation! Scramble(SVS) assembly.

The "V Rising" signature is the one actually matched RunFinalizer in SVS, so reorder this signature before the "Test Game" one.

Also add a comment for "Test Game" signature that matched on HoneyCome and DigitalCraft.